requirements.txt
elevenlabs
python-dotenv 